tratando mensagem de aluno deletado

+++ controllers/controller.go
@@ -47,6 +47,11 @@ func BuscaAlunoPorId(c *gin.Context) {
 func DeletaAluno(c *gin.Context) {
 	var aluno models.Aluno
 	id := c.Params.ByName("id")
+	database.DB.First(&aluno, id)
+	if aluno.ID == 0 {
+		c.JSON(http.StatusNotFound, gin.H{"Not found": "Aluno não encontrado"})
+		return
+	}
 	database.DB.Delete(&aluno, id)
 	c.JSON(http.StatusOK, gin.H{"data": "Aluno deletado com sucesso"})
 }
